feat: split built-in vs third-party abilities on one Connection screen
Agent Connector does two things: it runs an MCP server that exposes every
ability registered on the site (third-party abilities), and it ships its own
powerful abilities (shell, PHP eval, filesystem, WP-CLI, env-inspect,
admin-login). Those two are now separated and the admin UI is unified.

- Built-in abilities are hidden behind their own toggle, OFF by default
  (Config::builtin_abilities_enabled/active + BUILTIN_ABILITIES_OPTION). The
  master Enable toggle still runs the MCP server and exposes third-party
  abilities (always active while enabled); the built-in ones register only when
  their toggle is on.
- Settings and Connect collapse into a single top-level "Connection" screen
  (new Admin/ConnectionPage) with: status, Enable, production override,
  built-in-abilities toggle (with its own danger note), a third-party status
  line, domain lock + reconnect, and the application-password/connection
  generator. SettingsPage and ConnectPage are removed.
- Main file wakes the MCP adapter when can_boot(), and registers the built-in
  abilities (Plugin::register) only when the built-in toggle is on.
- Docs updated for the new model.
